+++ packages/zenrouter/lib/zenrouter.dart
@@ -18,6 +18,7 @@ export 'src/path/restoration.dart';
 export 'src/path/indexed.dart';
 export 'src/path/stack.dart';
 export 'src/path/transition.dart';
+export 'src/path/transition_delegate.dart';
 
 /// Route target capabilities
 export 'src/mixin/layout.dart';